Stable Debug Version
- Migrations are without any error
- Front end working.

+++ database/migrations/2021_08_09_052534_create_donations_table.php
@@ -17,7 +17,7 @@ class CreateDonationsTable extends Migration
             $table->id();
             $table->foreignId('campaign_id')->constrained();
             $table->bigInteger('payment_method_id');
-            $table->decimal('amount', 8, 2);
+            $table->double('amount', 8, 2);
             $table->bigInteger('donor_id');
             $table->string('donator_name');
             $table->string('note');
